mise en place de la carte de france

+++ Jeu_multiquiz/Point.pde
@@ -1,15 +1,19 @@
 class Point{
-  int x,y;
-  Point(int cx,int cy){
+  float x, y;
+  
+  Point(float cx, float cy){
     this.x = cx;
     this.y = cy;
   }
+  
   Point(){
     this.x = 0;
     this.y = 0;
   }
-  void setX(int cx){this.x=cx;}
-  void setY(int cy){this.y=cy;}
-  int getX(){return x;}
-  int getY(){return y;}
+  
+  void setX(float cx){ this.x = cx; }
+  void setY(float cy){ this.y = cy; }
+  float getX(){ return x; }
+  float getY(){ return y; }
+  
 }
